Add files via upload
This package uses some assets from sci-fi city and the first projectiles package. If nothing else, you can get the enemy behavior script from this, but you will need to give it a projectile that will instantiate with some velocity. The enemy has a range of sight where it will move and aim towards the player and shoot at the player.